+++ src/Pages/OATEditorPage/OATEditorPage.tsx
@@ -3,13 +3,11 @@ import OATHeader from '../../Components/OATHeader/OATHeader';
 import OATModelList from '../../Components/OATModelList/OATModelList';
 import OATGraphViewer from '../../Components/OATGraphViewer/OATGraphViewer';
 import OATPropertyEditor from '../../Components/OATPropertyEditor/OATPropertyEditor';
-import OATImport from './Internal/OATImport';
 import { getEditorPageStyles } from './OATEditorPage.Styles';
 import {
     OATEditorPageReducer,
     defaultOATEditorState
 } from './OATEditorPage.state';
-import { SET_OAT_IS_JSON_UPLOADER_OPEN } from '../../Models/Constants/ActionTypes';
 import i18n from '../../i18n';
 
 const OATEditorPage = ({ theme }) => {
@@ -18,13 +16,6 @@ const OATEditorPage = ({ theme }) => {
         defaultOATEditorState
     );
     const EditorPageStyles = getEditorPageStyles();
-
-    const handleImportClick = () => {
-        dispatch({
-            type: SET_OAT_IS_JSON_UPLOADER_OPEN,
-            payload: !state.isJsonUploaderOpen
-        });
-    };
 
     const languages = Object.keys(i18n.options.resources).map((language) => {
         return {
@@ -37,10 +28,6 @@ const OATEditorPage = ({ theme }) => {
         <div className={EditorPageStyles.container}>
             <OATHeader
                 elements={state.elements.digitalTwinsModels}
-                onImportClick={handleImportClick}
-            />
-            <OATImport
-                isJsonUploaderOpen={state.isJsonUploaderOpen}
                 dispatch={dispatch}
             />
             <div
